+++ ListableUI/Tests/Fixtures/PresentationState+Fixtures.swift
@@ -88,4 +88,6 @@ final class ReorderingActionsDelegateMock : ReorderingActionsDelegate
     ) {}
     
     func endReorder(for item: AnyPresentationItemState, with result: ReorderingActions.Result) {}
+    
+    func accessibilityMove(item: AnyPresentationItemState, direction: ReorderingActions.AccessibilityMoveDirection) -> Bool { true }
 }